Refactor(be): `AuthUser` 어노테이션 처리 로직 정리
