--- backend/grabtable/src/main/java/edu/skku/grabtable/auth/service/AuthUserArgumentResolver.java
@@ -1,36 +1,0 @@
-package edu.skku.grabtable.auth.service;
-
-import edu.skku.grabtable.auth.annotation.AuthUser;
-import edu.skku.grabtable.common.util.AuthHeaderParser;
-import jakarta.servlet.http.HttpServletRequest;
-import lombok.RequiredArgsConstructor;
-import org.springframework.core.MethodParameter;
-import org.springframework.stereotype.Component;
-import org.springframework.web.bind.support.WebDataBinderFactory;
-import org.springframework.web.context.request.NativeWebRequest;
-import org.springframework.web.method.support.HandlerMethodArgumentResolver;
-import org.springframework.web.method.support.ModelAndViewContainer;
-
-@Component
-@RequiredArgsConstructor
-public class AuthUserArgumentResolver implements HandlerMethodArgumentResolver {
-
-    private final LoginService loginService;
-
-    @Override
-    public boolean supportsParameter(MethodParameter parameter) {
-        return parameter.getParameterAnnotation(AuthUser.class) != null;
-    }
-
-    @Override
-    public Object resolveArgument(
-            MethodParameter parameter,
-            ModelAndViewContainer mavContainer,
-            NativeWebRequest webRequest,
-            WebDataBinderFactory binderFactory
-    ) {
-        HttpServletRequest request = (HttpServletRequest) webRequest.getNativeRequest();
-        String tokenStr = AuthHeaderParser.parseTokenString(request);
-        return loginService.getUserByToken(tokenStr);
-    }
-}
